Create odd-even-without-conditionals
This program defines two helper functions, is_even() and is_odd(), which check whether a given number is even or odd, respectively. Then, in the main() function, it takes an integer input from the user, utilizes a dictionary to map the boolean result of is_even() to the corresponding string ("even" or "odd"), and prints the result.